Add snarl decomposition building, generic Serializable, and per-class cpp files
